+++ Backend/resources/js/scripts.js
