+++ Backend/routes/web.php
@@ -14,4 +14,6 @@ use Illuminate\Support\Facades\Route;
 */
 
 
-
+Route::get('/', function () {
+    return view('pages.mainpage');
+});
